Update-delete first commit

+++ Recipe/DTOs/StepDto.cs
@@ -5,6 +5,6 @@ namespace Recipe.DTOs
     public class StepDto
     {
         public required string Description { get; set; }
-        public required int Order{ get; set; }
+        public required byte Order{ get; set; }
     }
 }
